Bank/requirements.txt
Django==1.7.1
Pillow==2.7.0
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-autofixture==0.9.2
django-tinymce==1.5.3
django-toolbelt==0.0.1
djangorestframework==3.0.4
gunicorn==19.1.1
psycopg2==2.5.4
static3==0.5.1
whitenoise==1.0.6
wsgiref==0.1.2
django-colorful
pytz